:recycle: Move createBaseTransaction() and its interface to utils and adjust its usages
