+++ packages/lisk-transactions/src/transactions/1_register_second_passphrase.ts
@@ -22,12 +22,8 @@ import {
 	Status,
 	TransactionJSON,
 } from '../transaction_types';
-import { validator } from '../utils';
-import {
-	BaseTransaction,
-	createBaseTransaction,
-	TransactionResponse,
-} from './base';
+import { createBaseTransaction, validator } from '../utils';
+import { BaseTransaction, TransactionResponse } from './base';
 
 const TRANSACTION_SIGNATURE_TYPE = 1;
