+++ packages/lisk-transactions/src/transactions/2_register_delegate.ts
@@ -21,13 +21,15 @@ import {
 	Status,
 	TransactionJSON,
 } from '../transaction_types';
-import { validator } from '../utils';
+import {
+	createBaseTransaction,
+	CreateBaseTransactionInput,
+	validator,
+} from '../utils';
 import { isTypedObjectArrayWithKeys } from '../utils/validation';
 import {
 	Attributes,
 	BaseTransaction,
-	createBaseTransaction,
-	CreateBaseTransactionInput,
 	ENTITY_ACCOUNT,
 	EntityMap,
 	TransactionResponse,
